// types/supabase.ts
export interface definitions {
  papers: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /** Format: text */
    name?: string;
    /** Format: text */
    magazine?: string;
    /** Format: bigint */
    position?: number;
    /** Format: text */
    side?: string;
  };
}